"fix: пофоиксил большинство замечаний"
