+++ main-service/src/main/java/ru/practicum/categories/controllers/CategoryAdminController.java
@@ -3,21 +3,19 @@ package ru.practicum.categories.controllers;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import ru.practicum.categories.dto.CategoryDto;
-import ru.practicum.categories.service.CategoryAdminService;
+import ru.practicum.categories.service.CategoryService;
 
 import javax.validation.Valid;
 
 @Slf4j
-@Validated
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/admin/categories")
 public class CategoryAdminController {
 
-    private final CategoryAdminService service;
+    private final CategoryService service;
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
