--- main-service/src/main/java/ru/practicum/categories/service/CategoryAdminService.java
@@ -1,11 +1,0 @@
-package ru.practicum.categories.service;
-
-import ru.practicum.categories.dto.CategoryDto;
-
-public interface CategoryAdminService {
-    CategoryDto postCategory(CategoryDto categoryDto);
-
-    CategoryDto patchCategory(long catId, CategoryDto categoryDto);
-
-    void deleteCategory(long catId);
-}
