+++ main-service/src/main/java/ru/practicum/categories/service/CategoryService.java
@@ -8,4 +8,10 @@ public interface CategoryService {
     CategoryDto getById(long catId);
 
     List<CategoryDto> getAll(int from, int size);
+
+    CategoryDto postCategory(CategoryDto categoryDto);
+
+    CategoryDto patchCategory(long catId, CategoryDto categoryDto);
+
+    void deleteCategory(long catId);
 }
