+++ main-service/src/main/java/ru/practicum/categories/service/CategoryServiceImpl.java
@@ -3,19 +3,24 @@ package ru.practicum.categories.service;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.categories.dto.CategoryDto;
 import ru.practicum.categories.dto.CategoryMapper;
 import ru.practicum.categories.model.Category;
 import ru.practicum.categories.repository.CategoryRepository;
+import ru.practicum.events.repository.EventsRepository;
 import ru.practicum.exception.NotFoundException;
+import ru.practicum.exception.RequestException;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
 @AllArgsConstructor
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
+    private final EventsRepository eventsRepository;
 
     @Override
     public CategoryDto getById(long catId) {
@@ -30,5 +35,34 @@ public class CategoryServiceImpl implements CategoryService {
                 .getContent().stream()
                 .map(CategoryMapper::toCategoryDto)
                 .collect(Collectors.toList());
+    }
+
+    @Override
+    @Transactional
+    public CategoryDto postCategory(CategoryDto categoryDto) {
+        Category category = CategoryMapper.fromDto(categoryDto);
+        Category saveCategory = categoryRepository.save(category);
+        return CategoryMapper.toCategoryDto(saveCategory);
+    }
+
+    @Override
+    @Transactional
+    public CategoryDto patchCategory(long catId, CategoryDto categoryDto) {
+        Category category = categoryRepository.findById(catId)
+                .orElseThrow(() -> new NotFoundException(String.format("Category with id=%d not found", catId)));
+
+        Optional.ofNullable(categoryDto.getName()).ifPresent(category::setName);
+        Category saveCategory = categoryRepository.save(category);
+        return CategoryMapper.toCategoryDto(saveCategory);
+    }
+
+    @Override
+    @Transactional
+    public void deleteCategory(long catId) {
+        categoryRepository.findById(catId)
+                .orElseThrow(() -> new NotFoundException(String.format("Category with id=%d not found", catId)));
+        if (!eventsRepository.findAllByCategory_Id(catId).isEmpty())
+            throw new RequestException("Category shouldn't have any events");
+        categoryRepository.deleteById(catId);
     }
 }
