+++ main-service/src/main/java/ru/practicum/compilations/controllers/CompilationsAdminController.java
@@ -4,7 +4,6 @@ import javax.validation.Valid;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -16,21 +15,20 @@ import org.springframework.web.bind.annotation.RestController;
 import ru.practicum.compilations.dto.CompilationCreateDto;
 import ru.practicum.compilations.dto.CompilationsDto;
 import ru.practicum.compilations.dto.CompilationsUpdateDto;
-import ru.practicum.compilations.service.CompilationsAdminService;
+import ru.practicum.compilations.service.CompilationsService;
 
 @RestController
 @RequestMapping({"/admin/compilations"})
-@Validated
 @Slf4j
 @AllArgsConstructor
 public class CompilationsAdminController {
-    private final CompilationsAdminService service;
+    private final CompilationsService service;
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public CompilationsDto postCompilations(@RequestBody @Valid CompilationCreateDto dto) {
+    public CompilationsDto createCompilations(@RequestBody @Valid CompilationCreateDto dto) {
         log.info("\nPOST [http://localhost:8080/admin/compilations] : запрос на создание подборки событий {}\n", dto);
-        return this.service.postCompilations(dto);
+        return service.createCompilations(dto);
     }
 
     @DeleteMapping({"/{compId}"})
@@ -38,13 +36,13 @@ public class CompilationsAdminController {
     public void deleteCompilationsById(@PathVariable int compId) {
         log.info("\nDELETE [http://localhost:8080/admin/compilations/{}] : запрос на удаление подборки событий с ID {}\n",
                 compId, compId);
-        this.service.deleteCompilation((long)compId);
+        service.deleteCompilation(compId);
     }
 
     @PatchMapping({"/{compId}"})
-    public CompilationsDto patchCompilation(@PathVariable long compId, @RequestBody @Valid CompilationsUpdateDto dto) {
+    public CompilationsDto updateCompilation(@PathVariable long compId, @RequestBody @Valid CompilationsUpdateDto dto) {
         log.info("\nPATCH [http://localhost:8080/admin/compilations/{}] : запрос на обновление подборки событий {} с ID {}\n",
                 compId, dto, compId);
-        return this.service.patchCompilation(compId, dto);
+        return service.updateCompilation(compId, dto);
     }
 }
