+++ main-service/src/main/java/ru/practicum/compilations/controllers/CompilationsController.java
@@ -2,7 +2,6 @@ package ru.practicum.compilations.controllers;
 
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -15,7 +14,6 @@ import java.util.List;
 
 @RestController
 @RequestMapping({"/compilations"})
-@Validated
 @Slf4j
 @AllArgsConstructor
 public class CompilationsController {
@@ -24,12 +22,12 @@ public class CompilationsController {
     @GetMapping
     public List<CompilationsDto> getCompilationsOfEvent(@RequestParam(defaultValue = "false") boolean pinned, @RequestParam(defaultValue = "0") int from, @RequestParam(defaultValue = "10") int size) {
         log.info("\nGET [http://localhost:8080/compilations] : запрос на просмотр подборок событий\n");
-        return this.service.getCompilationsOfEvent(pinned, from, size);
+        return service.getCompilationsOfEvent(pinned, from, size);
     }
 
     @GetMapping({"/{compId}"})
     public CompilationsDto getCompilationsOfEventById(@PathVariable long compId) {
         log.info("\nGET [http://localhost:8080/compilations/{}] : запрос для подборки событий по ID {}\n", compId, compId);
-        return this.service.getCompilationsOfEventById(compId);
+        return service.getCompilationsOfEventById(compId);
     }
 }
