+++ main-service/src/main/java/ru/practicum/compilations/dto/CompilationsMapper.java
@@ -3,6 +3,8 @@ package ru.practicum.compilations.dto;
 import lombok.experimental.UtilityClass;
 import ru.practicum.compilations.model.Compilations;
 
+import java.util.ArrayList;
+
 @UtilityClass
 public class CompilationsMapper {
     public static CompilationsDto toDto(Compilations compilations) {
@@ -10,7 +12,7 @@ public class CompilationsMapper {
                 .id(compilations.getId())
                 .title(compilations.getTitle())
                 .pinned(compilations.getPinned())
-                .events(compilations.getEvents())
+                .events(new ArrayList<>(compilations.getEvents()))
                 .build();
     }
 }
