+++ main-service/src/main/java/ru/practicum/compilations/model/Compilations.java
@@ -4,7 +4,7 @@ import lombok.*;
 import ru.practicum.events.model.Event;
 
 import javax.persistence.*;
-import java.util.List;
+import java.util.Set;
 
 @Entity
 @Table(name = "compilations")
@@ -25,5 +25,5 @@ public class Compilations {
     @ManyToMany
     @JoinTable(name = "compilations_events", joinColumns = @JoinColumn(name = "compilation_id"),
             inverseJoinColumns = @JoinColumn(name = "event_id"))
-    private List<Event> events;
+    private Set<Event> events;
 }
