--- main-service/src/main/java/ru/practicum/events/service/interfaces/AdminEventService.java
@@ -1,20 +1,0 @@
-package ru.practicum.events.service.interfaces;
-
-import ru.practicum.events.dto.EventFullDto;
-import ru.practicum.events.dto.EventUpdateDto;
-import ru.practicum.events.model.State;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-public interface AdminEventService {
-    List<EventFullDto> getAllEventByParameters(List<Long> users,
-                                               List<State> states,
-                                               List<Long> categories,
-                                               LocalDateTime rangeStart,
-                                               LocalDateTime rangeEnd,
-                                               int from,
-                                               int size);
-
-    EventFullDto patchEventByAdmin(long eventId, EventUpdateDto dto);
-}
