+++ main-service/src/main/java/ru/practicum/requests/controllers/UserRequestController.java
@@ -2,22 +2,20 @@ package ru.practicum.requests.controllers;
 
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import ru.practicum.requests.dto.ParticipationRequestDto;
 import ru.practicum.requests.dto.PatchRequestDto;
 import ru.practicum.requests.dto.PatchResponseDto;
-import ru.practicum.requests.service.UserRequestService;
+import ru.practicum.requests.service.RequestService;
 
 import java.util.List;
 
 @RestController
 @RequestMapping({"/users/{userId}/events/{eventId}/requests"})
-@Validated
 @Slf4j
 @AllArgsConstructor
 public class UserRequestController {
-    private final UserRequestService service;
+    private final RequestService service;
 
     @GetMapping
     public List<ParticipationRequestDto> getParticipateRequestsByEventCurrentUser(@PathVariable long userId,
@@ -25,15 +23,15 @@ public class UserRequestController {
         log.info("\nGET [http://localhost:8080/users/{}/events/{}/requests] : " +
                 "запрос от инициатора {} события {} на просмотр запросов на " +
                 "участие в событии\n", userId, eventId, userId, eventId);
-        return this.service.getParticipateRequestsByEventCurrentUser(userId, eventId);
+        return service.getParticipateRequestsByEventCurrentUser(userId, eventId);
     }
 
     @PatchMapping
-    public PatchResponseDto patchParticipateRequestsByEventCurrentUser(@PathVariable long userId,
+    public PatchResponseDto updateParticipateRequestsByEventCurrentUser(@PathVariable long userId,
                                                                        @PathVariable long eventId,
                                                                        @RequestBody PatchRequestDto request) {
         log.info("\nPATCH [http://localhost:8080/users/{}/events/{}/requests] : " +
                 "запрос на обновление статусов запросов на участие в событии {}\n", userId, eventId, request);
-        return this.service.patchParticipateRequestsByEventCurrentUser(userId, eventId, request);
+        return service.updateParticipateRequestsByEventCurrentUser(userId, eventId, request);
     }
 }
