+++ main-service/src/main/java/ru/practicum/utils/Constants.java
@@ -5,6 +5,8 @@ import java.time.format.DateTimeFormatter;
 
 public class Constants {
     public static final LocalDateTime MIN_TIME = LocalDateTime.of(2001, 1, 1, 0, 0);
-    public static final LocalDateTime MAX_TIME = LocalDateTime.of(9000, 12, 30, 23, 59);
     public static final DateTimeFormatter DATE_TIME_FORMAT_PATTERN = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
+
+    private Constants() {
+    }
 }
